Modified firewall behaviour
The wall speed starts slow and gradually increases over 20 seconds, as does the player speed. The distances between generated gaps have been reduced (300 -> 200), and when a gap is generated past a boundary, it is no longer generated at the boundary, but at a random distance from it (this was to prevent chains of gaps at the boundaries).